skip method now accepts long, fixme fixed
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/branches/ProActive_DataSpaces@13274 28e8926c-6b08-0410-baaa-805c5e19b8d6

+++ src/Extra/org/objectweb/proactive/extra/vfsprovider/server/FileSystemServerImpl.java
@@ -94,8 +94,7 @@ public class FileSystemServerImpl implements FileSystemServer {
         final Stream instance = tryGetStreamOrWound(stream);
         synchronized (instance) {
             checkContainsStreamOrWound(stream);
-            // FIXME: do it better than casting, which is only a hotfix to make it compile
-            return instance.skip((int) bytes);
+            return instance.skip(bytes);
         }
     }
